Replace concrete test algorithm class with dynamic mock
